pyproject.toml
[project]
name = "lumiaflow-bot"
version = "0.1.0"
description = "Telegram-first AI image generation and editing platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.7.0,<4.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "sqlalchemy>=2.0.32,<3.0.0",
    "alembic>=1.13.2,<2.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "psycopg[binary]>=3.2.1,<4.0.0",
    "redis>=5.0.8,<6.0.0",
    "celery>=5.4.0,<6.0.0",
    "httpx>=0.27.0,<1.0.0",
    "pydantic>=2.8.0,<3.0.0",
    "pydantic-settings>=2.4.0,<3.0.0",
    "python-json-logger>=2.0.7,<3.0.0",
    "structlog>=24.4.0,<25.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0,<8.0.0",
    "sentry-sdk[fastapi]>=2.12.0,<3.0.0",
    "boto3>=1.35.0,<2.0.0",
    "orjson>=3.10.7,<4.0.0",
    "google-genai>=1.30.0,<2.0.0",
    "pillow>=10.4.0,<11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "respx>=0.21.1,<1.0.0",
    "ruff>=0.6.3,<1.0.0",
    "mypy>=1.11.1,<2.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
